[SQUASH] sdm660: Upsreamed kernel to v4.4.232 * upstreamed through patch from kernel.org * Added some governers     - BLUE_ACTIVE     - ZZMOOVE     - ALUCARD     - NIGHTMARE     - DARKNESSV5 * Added Fast charge support * New Defconfig * Imported some dtsi's from stock 429 kernel

+++ build.sh
@@ -1,15 +1,16 @@
 #!/bin/bash
 
-set -o errexit
-
-[[ $# -eq 1 ]] || exit 1
-
-DEVICE=$1
+read -p 'Enter device codename: ' DEVICE
+read -p 'Enter kernel upstreamed version: ' VERSION
 
 if [[ $DEVICE != X00T ]]; then
-    echo invalid device codename
+    echo Invalid Device codename. Edith build.sh according to your device codename.
+    echo STOP!
     exit 1
 fi
+echo ''
+echo Starting kernel build.
+echo ''
 
 TOP=$(realpath ../)
 
@@ -25,6 +26,11 @@ PATH="$TOP/tools/misc/linux-x86/lz4:$PATH"
 PATH="$TOP/tools/misc/linux-x86/dtc:$PATH"
 PATH="$TOP/tools/misc/linux-x86/libufdt:$PATH"
 export LD_LIBRARY_PATH="$TOP/tools/clang/host/linux-x86/clang-r353983c/lib64:$LD_LIBRARY_PATH"
+
+rm -rf "$TOP/Output/$DEVICE-kernel"
+rm -rf "$TOP/ZIPS/hmp/Image.gz-dtb"
+rm -rf "$TOP/ZIPS/Kernel_ZIP/HMP"
+rm -rf out
 
 make \
     clean \
@@ -51,13 +57,34 @@ make \
     CLANG_TRIPLE=aarch64-linux-gnu- \
     CROSS_COMPILE=aarch64-linux-android- \
     CROSS_COMPILE_ARM32=arm-linux-androideabi-
-    
- 
 
-rm -rf "$TOP/Output/$DEVICE-kernel"
-mkdir -p "$TOP/Output/$DEVICE-kernel"
-cp out/arch/arm64/boot/Image.gz-dtb "$TOP/Output/$DEVICE-kernel"
-
+FILE='out/arch/arm64/boot/Image.gz-dtb'
+DIR="$TOP/ZIPS"
+if [[ -f "$FILE" ]] 
+   then    
+   mkdir -p "$TOP/Output/$DEVICE-kernel"
+   if [[ ! -d "$DIR" ]] 
+      then
+      echo ZIPS Folder not found !
+      echo Image.gz-dtb file is saved in $TOP/Output/$DEVICE-kernel folder.
+      exit 1
+   else
+   mkdir -p "$TOP/ZIPS/Kernel_ZIP/HMP"
+   cp out/arch/arm64/boot/Image.gz-dtb "$TOP/Output/$DEVICE-kernel"
+   cd "$TOP/Output/$DEVICE-kernel"
+   cp Image.gz-dtb "$TOP/ZIPS/hmp"
+   cd "$TOP/ZIPS/hmp"
+   zip -r "$TOP/ZIPS/Kernel_ZIP/HMP/Disaster_HMP-v$VERSION.zip" *
+   echo ''
+   echo ''
+   echo 'Kernel Build Successful !!!!'
+   fi
+else
+echo ''
+echo ''
+echo 'Kernel build Unsuccessfull !!!!'
+echo ''
+fi
